Factor out database qlpack inference.
